% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_BallHall}
\alias{CVI_BallHall}
\title{Negated Ball-Hall Index

TODO: update this docstring

Within cluster sum of squares weighted by the cluster cardinality}
\usage{
CVI_BallHall(X, y, K)
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}
}
\value{
The computed index.
}
\description{
Negated Ball-Hall Index

TODO: update this docstring

Within cluster sum of squares weighted by the cluster cardinality
}
