% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_CalinskiHarabasz}
\alias{CVI_CalinskiHarabasz}
\title{The Calinski-Harabasz Cluster Validity Index (Variance Ratio Criterion)

TODO: update this docstring}
\usage{
CVI_CalinskiHarabasz(X, y, K)
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}
}
\value{
The computed index.
}
\description{
The Calinski-Harabasz Cluster Validity Index (Variance Ratio Criterion)

TODO: update this docstring
}
\references{
T. Calinski & J. Harabasz. A dendrite method for cluster analysis,
Communications in Statistics, 3(1), 1974, pp. 1-27,
doi:10.1080/03610927408827101.
}
