% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_DaviesBouldin}
\alias{CVI_DaviesBouldin}
\title{The Negated Davies-Bouldin Cluster Validity Index

TODO: update this docstring}
\usage{
CVI_DaviesBouldin(X, y, K)
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}
}
\value{
The computed index, which is the additive inverse (negation)
        of the original Davies-Bouldin score. Negation is for consistency
        with other indices -- the higher the score, the better.
}
\description{
The Negated Davies-Bouldin Cluster Validity Index

TODO: update this docstring
}
\references{
D.L. Davies, D.W. Bouldin,
A Cluster Separation Measure,
IEEE Transactions on Pattern Analysis and Machine Intelligence. PAMI-1 (2),
1979, pp. 224-227, doi:10.1109/TPAMI.1979.4766909
}
