% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_Dunn}
\alias{CVI_Dunn}
\title{Dunn's index for measuring the degree to which clusters are compact separated

TODO: update this docstring

The index is defined by Eq.(3) in (Dunn, 1973).}
\usage{
CVI_Dunn(X, y, K)
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}
}
\value{
The computed index.
}
\description{
Dunn's index for measuring the degree to which clusters are compact separated

TODO: update this docstring

The index is defined by Eq.(3) in (Dunn, 1973).
}
\references{
J.C. Dunn, A Fuzzy Relative of the ISODATA Process and Its Use in Detecting
Compact Well-Separated Clusters, Journal of Cybernetics 3(3), 1973,
pp. 32-57, doi:10.1080/01969727308546046.
}
