% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_DuNNOWA}
\alias{CVI_DuNNOWA}
\title{OWA-based Dunn-like Indices Based on Near Neighbours

TODO: update this docstring}
\usage{
CVI_DuNNOWA(X, y, K, M = 10L, owa_numerator = "Min", owa_denominator = "Max")
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}

\item{M}{number of nearest neighbours}
}
\value{
The computed index.
}
\description{
OWA-based Dunn-like Indices Based on Near Neighbours

TODO: update this docstring
}
