% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_Gamma}
\alias{CVI_Gamma}
\title{The Baker-Hubert Gamma Coefficient

TODO: update this docstring

Equal to the Goodman-Kruskal rank correlation measure
between the vector of all pairwise distances and the 0/1 indicator
function stating whether the corresponding pair of points belongs
to the same cluster (0) or not (1).

Gives a value between -1 and 1. (NC-ND)/(NC+ND),
where NC - number of concordant and ND - number of discordant pairs.

F.B. Baker, L.J. Hubert, Measuring the power of hierarchical cluster
analysis, Journal of the American Statistical Association 70(349), 1975,
pp. 31-38.}
\usage{
CVI_Gamma(X, y, K)
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}
}
\value{
The computed index.
}
\description{
The Baker-Hubert Gamma Coefficient

TODO: update this docstring

Equal to the Goodman-Kruskal rank correlation measure
between the vector of all pairwise distances and the 0/1 indicator
function stating whether the corresponding pair of points belongs
to the same cluster (0) or not (1).

Gives a value between -1 and 1. (NC-ND)/(NC+ND),
where NC - number of concordant and ND - number of discordant pairs.

F.B. Baker, L.J. Hubert, Measuring the power of hierarchical cluster
analysis, Journal of the American Statistical Association 70(349), 1975,
pp. 31-38.
}
