% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_GDunn}
\alias{CVI_GDunn}
\title{Generalised Dunn's index

TODO: update this docstring}
\usage{
CVI_GDunn(X, y, K, lowercaseDelta, uppercaseDelta)
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}
}
\value{
The computed index.
}
\description{
Generalised Dunn's index

TODO: update this docstring
}
\references{
TODO: update
}
