% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_Silhouette}
\alias{CVI_Silhouette}
\title{The Silhouette Coefficient

TODO: update this docstring}
\usage{
CVI_Silhouette(X, y, K)
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}
}
\value{
The computed index.
}
\description{
The Silhouette Coefficient

TODO: update this docstring
}
\references{
P.J. Rousseeuw, Silhouettes: a Graphical Aid to the Interpretation and
Validation of Cluster Analysis, Computational and Applied Mathematics 20,
1987, pp. 53-65, doi:10.1016/0377-0427(87)90125-7.
}
