% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_WCNN}
\alias{CVI_WCNN}
\title{Within-Cluster Nearest-Neighbours

TODO: update this docstring}
\usage{
CVI_WCNN(X, y, K, M = 10L)
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}

\item{M}{number of nearest neighbours}
}
\value{
The computed index.
}
\description{
Within-Cluster Nearest-Neighbours

TODO: update this docstring
}
