% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVI_WCSS}
\alias{CVI_WCSS}
\title{Negated Within-Cluster Sum of Squares

TODO: update this docstring

Objective function used, amongst others, in the k-means and
the Ward and Caliński&Harabasz algorithms.}
\usage{
CVI_WCSS(X, y, K)
}
\arguments{
\item{X}{data matrix of size n*d}

\item{y}{vector of n integer labels in [1, K], where `y[i]`
is the cluster id of the i-th point, `X[i,]`}

\item{K}{number of clusters, `max(y)`}
}
\value{
The computed index.
}
\description{
Negated Within-Cluster Sum of Squares

TODO: update this docstring

Objective function used, amongst others, in the k-means and
the Ward and Caliński&Harabasz algorithms.
}
