% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deoptim.R
\name{deoptim}
\alias{deoptim}
\title{The DEoptim algorithm}
\usage{
deoptim(
  data_matrix,
  k,
  labels = NULL,
  population_size = 10 * length(labels),
  control = list(itermax = 1000),
  CVI_ptr = NULL
)
}
\arguments{
\item{data_matrix}{a nxk matrix representing points to cluster}

\item{k}{a number of clusters}

\item{labels}{an optional label vector from which the algorithm will start}

\item{population_size}{population size, default is as in DEoptim algorithm: 10 times length of a gene vector. It is only used if labels are not NULL. To set population size with labels equal to NULL use NP parameter in control parameter.}

\item{control}{a list of control parameters given to PSO algorithm, most notably itermax}

\item{CVI_ptr}{optional cluster validity index calculator to optimize. When NULL, CalinskiHarabasz is created and used.}
}
\value{
Returns the same value as DEoptim
}
\description{
The DEoptim algorithm is used to maximize a given cluster validity index
}
\details{
TBD
}
\seealso{
Other CVI: 
\code{\link{deoptim_many_centroids}()},
\code{\link{pso_many_centroids}()},
\code{\link{pso}()}
}
\concept{CVI}
