% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deoptim_many_centroids.R
\name{deoptim_many_centroids}
\alias{deoptim_many_centroids}
\title{The Many Centroids DEoptim algorithm}
\usage{
deoptim_many_centroids(
  data_matrix,
  k,
  centroids_per_cluster = 5,
  labels = NULL,
  population_size = 10 * k * centroids_per_cluster * ncol(data_matrix),
  control = list(itermax = 1000),
  fn_scale = 1,
  CVI_ptr = NULL
)
}
\arguments{
\item{data_matrix}{a nxk matrix representing points to cluster}

\item{k}{a number of clusters}

\item{centroids_per_cluster}{numer of centroids for every cluster}

\item{labels}{an optional label vector from which the algorithm will start}

\item{population_size}{population size, default is as in DEoptim algorithm: 10 times length of a gene vector. It is only used if labels are not NULL. To set population size with labels equal to NULL use NP parameter in control parameter.}

\item{control}{a list of control parameters given to DEoptim algorithm, most notably maxit}

\item{fn_scale}{a number by which the optimizing function will be multiplied. Default equal to 1}

\item{CVI_ptr}{optional cluster validity index calculator to optimize. When NULL, CalinskiHarabasz is created and used.}
}
\value{
Returns the same value as psoptim.
}
\description{
The Many Centroids DEoptim algorithm is used to maximize a given cluster validity index
}
\details{
TBD
}
\seealso{
Other CVI: 
\code{\link{deoptim}()},
\code{\link{pso_many_centroids}()},
\code{\link{pso}()}
}
\concept{CVI}
