% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.CVI_improve}
\alias{.CVI_improve}
\title{(Tabu-like) (stochastic) hill climbing}
\usage{
.CVI_improve(
  cvi_ptr,
  y0,
  allow_revisit = FALSE,
  max_iter_with_no_improvement = 250L,
  max_iter = 10000L,
  max_samples = -1L,
  verbose = FALSE
)
}
\arguments{
\item{cvi_ptr}{pointer, see _CVI_create()}

\item{y0}{initial label vector}

\item{allow_revisit}{should a `tabu` list be maintained?}

\item{max_iter_with_no_improvement}{how many iterations
where there are no improvement in the fitness function
before we give up?}

\item{max_iter}{maximal number of iterations}

\item{max_samples}{if <= 0, then an exhaustive search of all the
neighbouring points is conveyed; otherwise, choose
next candidates at random}

\item{verbose}{print additional info on the console?}
}
\value{
see optim()
}
\description{
(Tabu-like) (stochastic) hill climbing
}
