% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.CVI_improve_turbo}
\alias{.CVI_improve_turbo}
\title{Tabu-like hill climbing from multiple initial points}
\usage{
.CVI_improve_turbo(
  cvi_ptr,
  Y0,
  max_iter_with_no_improvement = 250L,
  max_iter = 10000L,
  verbose = FALSE
)
}
\arguments{
\item{cvi_ptr}{pointer, see _CVI_create()}

\item{Y0}{set of initial label vectors}

\item{max_iter_with_no_improvement}{how many iterations
where there are no improvement in the fitness function
before we give up?}

\item{max_iter}{maximal number of iterations}

\item{verbose}{print additional info on the console?}
}
\value{
see optim()
}
\description{
an exhaustive search of all the neighbouring points is conveyed
}
