% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso_many_centroids.R
\name{pso_many_centroids}
\alias{pso_many_centroids}
\title{The Many Centroids PSO algorithm}
\usage{
pso_many_centroids(
  data_matrix,
  k,
  centroids_per_cluster = 5,
  labels = NULL,
  control = list(maxit = 10000, reltol = 0.2),
  CVI_ptr = NULL
)
}
\arguments{
\item{data_matrix}{a nxk matrix representing points to cluster}

\item{k}{a number of clusters}

\item{centroids_per_cluster}{numer of centroids for every cluster}

\item{labels}{an optional label vector from which the algorithm will start}

\item{control}{a list of control parameters given to PSO algorithm, most notably maxit}

\item{CVI_ptr}{optional cluster validity index calculator to optimize. When NULL, CalinskiHarabasz is created and used.}
}
\value{
Returns the same value as psoptim.
}
\description{
The Many Centroids PSO algorithm is used to maximize a given cluster validity index
}
\details{
TBD
}
\seealso{
Other CVI: 
\code{\link{deoptim_many_centroids}()},
\code{\link{deoptim}()},
\code{\link{pso}()}
}
\concept{CVI}
